anykernel: fix location of firmware files (OTG-USB Device fix!)
due to different default location of firmware files on OOS5 roms
it wasn't possible to load any provided usb firmware files because
they were placed in the wrong directory. the right directory is a
dedicated mounted 'firmware' directory which is on the modem partition.
files are placed there now and therefore can be loaded again.

thx to fredrickz for coming up with this fix